fix: improved code based on the review
