--- packages/api/src/constants/token.ts
@@ -1,9 +1,0 @@
-interface TokenObject {
-  OIDC_TOKEN: string;
-}
-
-export const TOKEN_KEY = "OIDC_TOKEN";
-
-export const TOKEN_OBJECT: TokenObject = {
-  [TOKEN_KEY]: "",
-};
