+++ packages/api/src/index.ts
@@ -1,6 +1,5 @@
 export { default as ApiContext, ApiProvider } from "./contexts/ApiContext";
 export * from "./constants/apollo-cache";
-export * from "./constants/token";
 export * from "./generated/Graphql";
 export * from "./queries/get-taken";
 export * from "./queries/get-taken-v2";
