+++ packages/authentication/package.json
@@ -28,7 +28,9 @@
   "dependencies": {
     "@gemeente-denhaag/modal": "^1.1.2",
     "@gemeente-denhaag/typography": "1.1.2",
-    "jwt-decode": "^4.0.0"
+    "jwt-decode": "^4.0.0",
+    "oidc-client-ts": "3.1.0",
+    "react-oidc-context": "^3.2.0"
   },
   "devDependencies": {
     "@types/jest": "^29.5.12",
@@ -42,11 +44,9 @@
     "vitest": "^2.0.4"
   },
   "peerDependencies": {
-    "oidc-client-ts": "3.1.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-intl": "^6.6.8",
-    "react-oidc-context": "^3.2.0",
     "react-router-dom": "^6.27.0"
   }
 }
