+++ packages/authentication/src/index.ts
@@ -1,5 +1,5 @@
-export { default as OidcWrapper } from "./components/OidcWrapper";
-export { default as OidcContext } from "./contexts/OidcContext";
+export { default as OidcProvider } from "./providers/OidcProvider";
+export * from "./providers/OidcProvider";
 export { default as OidcCallbackPage } from "./pages/OidcCallbackPage";
 export { default as ProtectedApp } from "./components/ProtectedApp";
 export { default as useLogout } from "./hooks/useLogout";
