--- packages/authentication/src/tests/components/KeycloakWrapper.test.tsx
@@ -1,7 +1,0 @@
-import OidcWrapper from "../../components/OidcWrapper";
-
-describe("OidcWrapper", () => {
-  it("is truthy", () => {
-    expect(OidcWrapper).toBeTruthy();
-  });
-});
