+++ packages/authentication/src/tests/providers/OidcProvider.test.tsx
@@ -1,0 +1,7 @@
+import OidcProvider from "../../providers/OidcProvider";
+
+describe("OidcProvider", () => {
+  it("is truthy", () => {
+    expect(OidcProvider).toBeTruthy();
+  });
+});
